+++ config/babel.config.js
