+++ lib/devServer.js
@@ -5,8 +5,8 @@ var express = require('express')
 var compress = require('compression')
 var app = express()
 
-var hipley = require('./')
-var config = require('./webpack.config')({
+var hipley = require('../')
+var config = require('../config/webpack.config')({
   development: true
 })
 var compiler = webpack(config)
