+++ lib/server.js
@@ -2,7 +2,7 @@ var path = require('path')
 var compress = require('compression')
 var express = require('express')
 var app = express()
-var hipley = require('./')
+var hipley = require('../')
 
 var ROOT = hipley.options.root
 var DEST = path.resolve(ROOT, hipley.options.dest)
